// vessel/src/app/protected/interfaces/mongoInter.ts
export interface Vessel {
  _id:                    string;
  Bale:                   string;
  Beam_m:                 string;
  Builder:                string;
  Classification_Society: string;
  Crude_Oil_bbl:          string;
  Distance__Time:         string;
  Draught_m:              string;
  Flag:                   string;
  Gas_m3:                 string;
  Grain:                  string;
  Gross_Tonnage:          string;
  Homeport:               string;
  IMO_number:             string;
  Lat:                    number;
  Length_Overall_m:       string;
  Lon:                    number;
  Manager:                string;
  Photo:                  string;
  Place_of_Built:         string;
  Predicted_ETA:          string;
  Registered_Owner:       string;
  Ship_type:              string;
  Source:                 string;
  Summer_Deadweight_t:    string;
  TEU:                    string;
  Vessel_Name:            string;
  Yard:                   string;
  Year_of_Built:          string;
}